Separate many classes into isolated files
