+++ include/Hexagon/ClipboardData.hpp
@@ -1,0 +1,18 @@
+#pragma once
+
+
+
+#include <string>
+#include <vector>
+
+
+
+class ClipboardData
+{
+public:
+   static bool store(std::vector<std::string> &lines);
+   static std::vector<std::string> retrieve();
+};
+
+
+
