+++ include/Hexagon/CodeRange.hpp
@@ -1,0 +1,35 @@
+#pragma once
+
+
+
+#include <Hexagon/CodePoint.hpp>
+
+
+
+class CodeRange
+{
+private:
+   int cursor_anchor_x;
+   int cursor_anchor_y;
+   int cursor_end_x;
+   int cursor_end_y;
+
+public:
+   CodeRange(int cursor_anchor_x, int cursor_anchor_y, int cursor_end_x, int cursor_end_y);
+   ~CodeRange();
+
+   CodePoint infer_cursor_start();
+   CodePoint infer_cursor_end();
+   void set_cursor_end_x(int x);
+   void set_cursor_end_y(int y);
+   bool is_empty();
+   int infer_num_lines();
+   bool in_range(int x, int y);
+};
+
+
+
+std::ostream &operator<<(std::ostream &out, CodeRange &code_range);
+
+
+
