+++ include/Hexagon/shared_globals.hpp
@@ -1,0 +1,15 @@
+#pragma once
+
+
+
+#include <string>
+
+
+
+extern std::string REGEX_TEMP_FILENAME;
+extern std::string CLIPBOARD_TEMP_FILENAME;
+extern std::string FILE_NAVIGATOR_SELECTION_FILENAME;
+extern std::string FILE_NAVIGATOR_SELECTION_last_content;
+
+
+
