+++ programs/hexagon.cpp
@@ -34,10 +34,11 @@ using namespace Hexagon;
 
 
 
-std::string REGEX_TEMP_FILENAME = "regex.txt";
-std::string CLIPBOARD_TEMP_FILENAME = "clipboard.txt";
-std::string FILE_NAVIGATOR_SELECTION_FILENAME = "file_navigator_selection.txt";
-std::string FILE_NAVIGATOR_SELECTION_last_content = "";
+#include <Hexagon/shared_globals.hpp>
+//std::string REGEX_TEMP_FILENAME = "regex.txt";
+//std::string CLIPBOARD_TEMP_FILENAME = "clipboard.txt";
+//std::string FILE_NAVIGATOR_SELECTION_FILENAME = "file_navigator_selection.txt";
+//std::string FILE_NAVIGATOR_SELECTION_last_content = "";
 
 
 
@@ -58,315 +59,25 @@ std::string resource(std::vector<std::string> components, std::string filename="
 
 
 
-std::string::size_type rfind_first_not_of(const std::string& s,
-        const std::string& delims, std::string::size_type pos)
-{
-    std::string::size_type p = pos;
-    while (delims.find(s[p]) != std::string::npos) {
-        if (p == 0) // not found
-            return std::string::npos;
-        --p;
-    }
-    return p;
-}
-
-std::string::size_type rfind_first_of(const std::string& s,
-        const std::string& delims, std::string::size_type pos)
-{
-    std::string::size_type p = pos;
-    while (delims.find(s[p]) == std::string::npos) {
-        if (p == 0) // not found
-            return std::string::npos;
-        --p;
-    }
-    return p;
-}
-
-#include <iostream>       // std::cout
-#include <string>         // std::string
-#include <cstddef>        // std::size_t
-
-int find_whitespace(std::string &string, int current_cursor_position)
-{
-  std::size_t position = string.find_first_of(" \n\r\t", current_cursor_position);
-  if (position != std::string::npos) return position;
-  return 0;
-}
-
-int find_non_whitespace(std::string &string, int current_cursor_position)
-{
-  std::size_t position = string.find_first_not_of(" \n\r\t", current_cursor_position);
-  if (position != std::string::npos) return position;
-  return 0;
-}
-
-int rfind_whitespace(std::string &string, int current_cursor_position)
-{
-  std::size_t position = rfind_first_of(string, " \n\r\t", current_cursor_position);
-  if (position != std::string::npos) return position;
-  return 0;
-}
-
-int rfind_non_whitespace(std::string &string, int current_cursor_position)
-{
-  std::size_t position = rfind_first_not_of(string, " \n\r\t", current_cursor_position);
-  if (position != std::string::npos) return position;
-  return 0;
-}
-
-
-#include <fstream>
-
-bool read_file(std::vector<std::string> &lines, std::string filename)
-{
-   std::ifstream file(filename);
-
-   lines.clear();
-
-   if (file.is_open())
-   {
-      std::string line;
-      while (getline(file, line)) { lines.push_back(line); }
-      file.close();
-   }
-
-   return true;
-}
-
-bool save_file(std::vector<std::string> &lines, std::string filename)
-{
-   std::ofstream file(filename);
-
-   if (file.is_open())
-   {
-      for (auto &line : lines) file << line << '\n';
-      file.close();
-   }
-   else
-   {
-      std::stringstream error_message;
-      error_message << "save_file(): Couldn't save file: file failed at is_open() on the filename \"" << filename << "\"" << std::endl;
-      throw std::runtime_error(error_message.str());
-   }
-
-   return true;
-}
+#include <Hexagon/util.hpp>
 
 
 
-class ClipboardData
-{
-public:
-   static bool store(std::vector<std::string> &lines)
-   {
-      ::save_file(lines, CLIPBOARD_TEMP_FILENAME);
-      return true;
-   }
-   static std::vector<std::string> retrieve()
-   {
-      std::vector<std::string> results;
-      ::read_file(results, CLIPBOARD_TEMP_FILENAME);
-      return results;
-   }
-};
+#include <Hexagon/ClipboardData.hpp>
 
 
 
-class ActionData
-{
-private:
-   enum datatype {
-      UNBOUND,
-      STRING,
-      CHAR,
-      BOOL,
-      FLOAT,
-      DOUBLE,
-      INT,
-   };
+#include <Hexagon/ActionData.hpp>
 
-   datatype type;
-   const std::string ATTRIBUTE_KEY = "value";
 
-   static std::string _get_datatype_string(datatype type)
-   {
-      switch(type)
-      {
-         case UNBOUND:
-            return "unbound";
-            break;
-         case STRING:
-            return "string";
-            break;
-         case CHAR:
-            return "char";
-            break;
-         case BOOL:
-            return "bool";
-            break;
-         case FLOAT:
-            return "float";
-            break;
-         case DOUBLE:
-            return "double";
-            break;
-         case INT:
-            return "int";
-            break;
-         default:
-            throw std::runtime_error("[ActionData]: Unidentified attribute type - boom!");
-            break;
-      };
-   }
-
-   std::string string_value;
-
-public:
-   ActionData() : type(UNBOUND) {}
-   ActionData(std::string string_value) : type(STRING), string_value(string_value) {}
-
-   ~ActionData() {}
-
-   std::string get_string()
-   {
-      if (type != STRING)
-      {
-         std::stringstream error_message;
-         error_message << "Attempting to get a string attribute when it is defined as another type \"" << _get_datatype_string(type) << "\"." << std::endl;
-         throw std::runtime_error(error_message.str());
-      }
-      return string_value;
-   }
-};
+#include <Hexagon/Action.hpp>
 
 
 
-class Action
-{
-private:
-   std::string name;
-   ActionData data1;
+#include <Hexagon/ActionQueueRecording.hpp>
 
-public:
-   Action(std::string name, ActionData data1)
-      : name(name)
-      , data1(data1)
-   {}
-
-   ~Action() {}
-
-   std::string get_name() { return name; }
-   ActionData get_data1() { return data1; }
-};
-
-
-
-class ActionQueueRecording
-{
-private:
-   std::string name;
-   std::vector<Action> actions;
-
-public:
-   ActionQueueRecording(std::string name)
-      : name(name)
-   {}
-
-   void clear_actions() { actions.clear(); }
-   std::string get_name() { return name; }
-   void append_action(Action action) { actions.push_back(action); }
-   int infer_num_actions() { return actions.size(); }
-   Action get_action_at(int index)
-   {
-      if (index < 0 || index >= actions.size())
-      {
-         std::stringstream error_message;
-         error_message << "\033[0;31m" << "[ActionQueueRecording]: Error attempting to access index ("
-            << index << ") which is outside of range (0-" << actions.size() << ")" << "\033[0m" << std::endl;
-         throw std::runtime_error(error_message.str());
-      }
-
-      return actions[index];
-   }
-};
-
-class GitLinesModifiedExtractor
-{
-private:
-   std::string filename;
-   std::vector<int> lines_modified;
-   bool executed;
-
-public:
-   GitLinesModifiedExtractor(std::string filename)
-      : filename(filename)
-      , lines_modified()
-      , executed(false)
-   {}
-   ~GitLinesModifiedExtractor()
-   {
-   }
-
-   void execute()
-   {
-      if (executed) return;
-
-      std::stringstream git_changed_lines_shell_command;
-      git_changed_lines_shell_command << "git blame " << filename << " | grep -n '^0\\{8\\} ' | cut -f1 -d:";
-      ShellCommandExecutor shell_command_executor(git_changed_lines_shell_command.str());
-      std::string output = shell_command_executor.execute();
-      std::vector<std::string> line_numbers_str = StringSplitter(output, '\n').split();
-      lines_modified.clear();
-      lines_modified.resize(line_numbers_str.size());
-      for (auto &line_number_str : line_numbers_str)
-         lines_modified.push_back(atoi(line_number_str.c_str()));
-      executed = true;
-   }
-
-   std::vector<int> get_lines_modified()
-   {
-      return lines_modified;
-   }
-};
-
-
-
-class CodePoint
-{
-private:
-   int x;
-   int y;
-
-public:
-   CodePoint(int x, int y)
-      : x(x)
-      , y(y)
-   {}
-
-   int get_x() const { return x; }
-   int get_y() const { return y; }
-   void set_x(int x) { this->x = x; }
-   void set_y(int y) { this->y = y; }
-};
-
-
-
-bool operator<(const CodePoint &a, const CodePoint &other)
-{
-   if (a.get_y() < other.get_y()) return true;
-   if (a.get_y() > other.get_y()) return false;
-   if (a.get_y() == other.get_y()) return (a.get_x() < other.get_x());
-   throw std::runtime_error("Codepoint operator<: unexpected codepath");
-}
-
-
-
-std::ostream &operator<<(std::ostream &out, CodePoint &code_point)
-{
-   out << "(" << code_point.get_x() << ", " << code_point.get_y() << ")" << std::endl;
-   return out;
-}
-
+#include <Hexagon/CodePoint.hpp>
+#include <Hexagon/CodeRange.hpp>
 
 
 class RailsMinitestTestResult
@@ -620,375 +331,11 @@ namespace CppCompiler
 
 
 
+#include <Hexagon/CodeRangeExtractor.hpp>
 
 
-class CodeRange
-{
-private:
-   int cursor_anchor_x;
-   int cursor_anchor_y;
-   int cursor_end_x;
-   int cursor_end_y;
+#include <Hexagon/CodeMessagePoint.hpp>
 
-public:
-   CodeRange(int cursor_anchor_x, int cursor_anchor_y, int cursor_end_x, int cursor_end_y)
-      : cursor_anchor_x(cursor_anchor_x)
-      , cursor_anchor_y(cursor_anchor_y)
-      , cursor_end_x(cursor_end_x)
-      , cursor_end_y(cursor_end_y)
-   {}
-   ~CodeRange() {}
-
-   // attributes
-
-   //int get_cursor_anchor_x() { return cursor_anchor_x; }
-   //int get_cursor_anchor_y() { return cursor_anchor_y; }
-   //int get_cursor_end_x() { return cursor_end_x; }
-   //int get_cursor_end_y() { return cursor_end_y; }
-
-   CodePoint infer_cursor_start()
-   {
-      return std::min(CodePoint(cursor_anchor_x, cursor_anchor_y), CodePoint(cursor_end_x, cursor_end_y));
-   }
-
-   CodePoint infer_cursor_end()
-   {
-      return std::max(CodePoint(cursor_anchor_y, cursor_anchor_y), CodePoint(cursor_end_x, cursor_end_y));
-   }
-
-   void set_cursor_end_x(int x) { cursor_end_x = x; }
-   void set_cursor_end_y(int y) { cursor_end_y = y; }
-
-   // inference
-
-   bool is_empty()
-   {
-      return cursor_end_y == cursor_anchor_y && cursor_end_x == cursor_anchor_x;
-   }
-
-   int infer_num_lines()
-   {
-      if (is_empty()) return 0;
-      CodePoint start = infer_cursor_start();
-      CodePoint end = infer_cursor_end();
-
-      return end.get_y() - start.get_y() + 1;
-   }
-
-   bool in_range(int x, int y)
-   {
-     // if at start, in range
-     if (y < cursor_anchor_y) return false;
-     else if (y > cursor_end_y) return false;
-     else if (y == cursor_anchor_y == cursor_end_y)
-     {
-        if (x < cursor_anchor_x) return false;
-        if (x >= cursor_end_x) return false;
-        return true;
-     }
-
-     throw std::runtime_error("Hmm, the code never should have gotten here.  There's an error in the logic");
-   }
-};
-
-
-
-std::ostream &operator<<(std::ostream &out, CodeRange &code_range)
-{
-   CodePoint start = code_range.infer_cursor_start();
-   CodePoint end = code_range.infer_cursor_end();
-   out << "(" << start.get_x() << ", " << start.get_y() << ")-(" << end.get_x() << ", " << end.get_y() << ")" << std::endl;
-   return out;
-}
-
-
-
-class CodeRangeRenderer
-{
-private:
-   const std::vector<std::string> &lines;
-   CodeRange &code_range;
-   int first_line;
-   int cell_width;
-   int cell_height;
-
-public:
-   CodeRangeRenderer(const std::vector<std::string> &lines, CodeRange &code_range, int first_line, int cell_width, int cell_height)
-      : lines(lines)
-      , code_range(code_range)
-      , first_line(first_line)
-      , cell_width(cell_width)
-      , cell_height(cell_height)
-   {}
-
-   bool verify_line_in_range(int line_num)
-   {
-      if (line_num >= lines.size()) return false;
-      if (line_num < 0) return false;
-      return true;
-   }
-
-   int get_line_length(int line_num)
-   {
-      if (!verify_line_in_range(line_num)) return 0;
-      return lines[line_num].length();
-   }
-
-   void render()
-   {
-      int num_lines = code_range.infer_num_lines();
-      CodePoint start = code_range.infer_cursor_start();
-      CodePoint end = code_range.infer_cursor_end();
-
-      ALLEGRO_COLOR selection_color = al_color_name("orange");
-      selection_color.r *= 0.4;
-      selection_color.g *= 0.4;
-      selection_color.b *= 0.4;
-      selection_color.a *= 0.4;
-
-      if (num_lines == 0) return;
-      if (num_lines == 1)
-      {
-         // draw beginning-to-end on single line
-         //CodePoint start = code_range.infer_cursor_start();
-         //CodePoint end = code_range.infer_cursor_end();
-
-         int this_actual_line_y = start.get_y();
-         int this_line_y = (this_actual_line_y - first_line);
-
-         al_draw_filled_rectangle(
-            start.get_x() * cell_width,
-            this_line_y * cell_height,
-            start.get_x() * cell_width,
-            (this_line_y + 1) * cell_height,
-            selection_color
-         );
-      }
-      if (num_lines >= 2)
-      {
-         // draw first line
-         //CodePoint start = code_range.infer_cursor_start();
-         //CodePoint end = code_range.infer_cursor_end();
-
-         int this_actual_line_y = start.get_y();
-         int this_line_y = this_actual_line_y - first_line;
-
-         al_draw_filled_rectangle(
-            start.get_x() * cell_width,
-            this_line_y * cell_height,
-            get_line_length(this_actual_line_y) * cell_width,
-            (this_line_y + 1) * cell_height,
-            selection_color
-         );
-
-         if (num_lines > 2)
-         {
-            // draw intermediate lines
-            //CodePoint start = code_range.infer_cursor_start();
-            //CodePoint end = code_range.infer_cursor_end();
-
-            for (int i = (start.get_y()+1); i < end.get_y(); i++)
-            {
-               this_line_y = (i - first_line);
-               al_draw_filled_rectangle(0, this_line_y * cell_height, get_line_length(i) * cell_width, (this_line_y + 1) * cell_height, selection_color);
-            }
-         }
-
-         // draw last line
-
-         //CodePoint start = code_range.infer_cursor_start();
-         //CodePoint end = code_range.infer_cursor_end();
-
-         this_actual_line_y = end.get_y();
-         this_line_y = (this_actual_line_y - first_line);
-         al_draw_filled_rectangle(0, (this_line_y * cell_height), end.get_x() * cell_width, (this_line_y+1) * cell_height, selection_color);
-      }
-   }
-};
-
-
-
-class CodeMessagePoint
-{
-public:
-   enum type_t
-   {
-      NONE,
-      ERROR,
-      TEST_FAILURE,
-      POSITION,
-   };
-
-private:
-   int x, y, length;
-   int cursor_placement_offset;
-   std::string message;
-   type_t type;
-
-public:
-   CodeMessagePoint(int x, int y, int length, int cursor_placement_offset, std::string message, type_t type)
-      : x(x)
-      , y(y)
-      , length(length)
-      , cursor_placement_offset(cursor_placement_offset)
-      , message(message)
-      , type(type)
-   {}
-   ~CodeMessagePoint()
-   {}
-
-   int get_x() { return x; }
-   int get_x2() { return x + length; }
-   int get_y() { return y; }
-   int get_length() { return length; }
-   int get_cursor_placement_offset() { return cursor_placement_offset; }
-   std::string get_message() { return message; }
-   type_t get_type() { return type; }
-
-   bool infer_at_point(int x, int y)
-   {
-      return (x == this->x && y == this->y);
-   }
-};
-
-
-
-class CodeMessagePointRenderer
-{
-private:
-   CodeMessagePoint code_message_point;
-   int current_line_number_offset;
-   int line_height;
-   int character_width;
-   ALLEGRO_FONT *font;
-   placement2d place;
-   int cursor_x;
-   int cursor_y;
-
-public:
-   CodeMessagePointRenderer(CodeMessagePoint code_message_point, ALLEGRO_FONT *font, int current_line_number_offset, int line_height, int character_width, int cursor_x, int cursor_y)
-      : code_message_point(code_message_point)
-      , current_line_number_offset(current_line_number_offset)
-      , line_height(line_height)
-      , character_width(character_width)
-      , font(font)
-      , place(0, 0, 2200, 500)
-      , cursor_x(cursor_x)
-      , cursor_y(cursor_y)
-   {
-      if (!font) throw std::runtime_error("font not valid in CodeMessagePointRenderer");
-   }
-
-   void render()
-   {
-      float horizontal_padding = 20;
-      place.position = vec2d(code_message_point.get_x() * character_width, (code_message_point.get_y()-1)*line_height - (current_line_number_offset)*line_height + line_height*0.5);
-      place.align = vec2d(0, 0);
-      place.start_transform();
-      ALLEGRO_COLOR spot_color, background_color, text_color, outline_color;
-
-      switch(code_message_point.get_type())
-      {
-      case CodeMessagePoint::POSITION:
-         {
-            spot_color = al_color_name("dodgerblue");
-            outline_color = al_color_name("yellow");
-            ALLEGRO_STATE blender_state;
-            float padding_x = 2;
-            float padding_y = 2;
-            float radius = 3;
-            float outline_thickness = 2.0;
-            bool cursor_at_code_point = ((cursor_x >= code_message_point.get_x()) && (cursor_x < code_message_point.get_x2()) && (code_message_point.get_y()-1) == cursor_y);
-            if (!cursor_at_code_point)
-            {
-               float opacity = 0.6;
-               spot_color.r *= opacity; spot_color.g *= opacity, spot_color.b *= opacity; spot_color.a *= opacity;
-            }
-            al_store_state(&blender_state, ALLEGRO_STATE_BLENDER);
-            al_set_blender(ALLEGRO_ADD, ALLEGRO_ONE, ALLEGRO_ONE);
-            al_draw_filled_rounded_rectangle(0-padding_x, -line_height/2.0-padding_y, std::max(10, code_message_point.get_length()*character_width)+padding_x, line_height/2.0+padding_y, radius, radius, spot_color);
-            if (cursor_at_code_point) al_draw_rounded_rectangle(0-padding_x, -line_height/2.0-padding_y, std::max(10, code_message_point.get_length()*character_width)+padding_x, line_height/2.0+padding_y, radius, radius, outline_color, outline_thickness);
-            al_restore_state(&blender_state);
-         }
-         break;
-      case CodeMessagePoint::ERROR:
-         {
-            spot_color = al_color_name("red");
-            background_color = al_color_name("webmaroon");
-            text_color = al_color_name("red");
-            float radius = 6.0;
-            float padding = 6.0;
-
-            // draw the marker
-            al_draw_filled_circle(0, 0, 16, spot_color);
-
-            /// draw the message (if open)
-            if (code_message_point.infer_at_point(cursor_x, cursor_y+1))
-            {
-               al_draw_filled_rounded_rectangle(-padding*2, -padding*2, place.size.x+padding*2, place.size.y+padding*2, radius, radius, al_color_name("black"));
-               al_draw_filled_rectangle(0, 0, place.size.x, place.size.y, background_color);
-               al_draw_rounded_rectangle(-padding, -padding, place.size.x+padding, place.size.y+padding, radius, radius, text_color, 3.0);
-               al_draw_multiline_text(font, text_color, 10, 10, place.size.x - horizontal_padding*2, line_height, ALLEGRO_ALIGN_LEFT, code_message_point.get_message().c_str());
-            }
-         }
-         break;
-      case CodeMessagePoint::TEST_FAILURE:
-      default:
-         {
-            spot_color = al_color_name("deeppink");
-            background_color = al_color_name("webmaroon");
-            text_color = al_color_name("deeppink");
-            float radius = 6.0;
-            float padding = 6.0;
-
-            // draw the marker
-            al_draw_filled_circle(0, 0, 16, spot_color);
-
-            /// draw the message (if open)
-            if (code_message_point.infer_at_point(cursor_x, cursor_y+1))
-            {
-               al_draw_filled_rounded_rectangle(-padding*2, -padding*2, place.size.x+padding*2, place.size.y+padding*2, radius, radius, al_color_name("black"));
-               al_draw_filled_rectangle(0, 0, place.size.x, place.size.y, background_color);
-               al_draw_rounded_rectangle(-padding, -padding, place.size.x+padding, place.size.y+padding, radius, radius, text_color, 3.0);
-               al_draw_multiline_text(font, text_color, 10, 10, place.size.x - horizontal_padding*2, line_height, ALLEGRO_ALIGN_LEFT, code_message_point.get_message().c_str());
-            }
-         }
-         break;
-      }
-
-      place.restore_transform();
-   }
-};
-
-
-
-class CodeRangeExtractor
-{
-private:
-   const std::vector<std::string> &lines;
-   CodeRange &code_range;
-
-public:
-   CodeRangeExtractor(const std::vector<std::string> &lines, CodeRange &code_range)
-      : lines(lines)
-      , code_range(code_range)
-   {}
-
-   std::vector<std::string> extract()
-   {
-      CodePoint start = code_range.infer_cursor_start();
-      CodePoint end = code_range.infer_cursor_end();
-
-      int start_y = std::min(std::max(0, start.get_y()), (int)lines.size());
-      int end_y = std::min(std::max(0, (end.get_y()+1)), (int)lines.size());
-
-      std::cout << "CodeRangeExtractor: attempting extraction at (" << start_y << ", " << end_y << ")" << std::endl;
-
-      std::vector<std::string> result(lines.begin() + start_y, lines.begin() + end_y);
-
-      return result;
-   }
-};
 
 
 
@@ -1013,32 +360,10 @@ public:
 
 
 
-class CodeMessagePointsOverlay
-{
-public:
-   std::string name;
-   ALLEGRO_COLOR color;
-   std::vector<CodeMessagePoint> code_message_points;
 
-   CodeMessagePointsOverlay(ALLEGRO_COLOR color, std::vector<CodeMessagePoint> code_message_points)
-      : color(color)
-      , code_message_points(code_message_points)
-   {
-      float alpha = 0.2;
-      color.r *= alpha; color.g *= alpha; color.b *= alpha; color.a *= alpha;
-   }
+#include <Hexagon/CodeMessagePointsOverlay.hpp>
 
-   void render(ALLEGRO_FONT *font, int first_line_number, float line_height, int cursor_x, int cursor_y)
-   {
-      //al_draw_filled_rectangle(0, 0, 2400, 2000, color);
-      int character_width = al_get_text_width(font, "x");
-      for (auto &code_message_point : code_message_points)
-      {
-         CodeMessagePointRenderer code_message_point_renderer(code_message_point, font, first_line_number, al_get_font_line_height(font), character_width, cursor_x, cursor_y);
-         code_message_point_renderer.render();
-      }
-   }
-};
+
 
 
 
@@ -1051,56 +376,7 @@ SystemWideMotion motion;
 
 
 
-
-
-class StageInterface
-{
-public:
-   enum type_t
-   {
-      NONE,
-      CODE_EDITOR,
-      ONE_LINE_INPUT_BOX,
-      FILE_NAVIGATOR,
-      KEYBOARD_INPUTS_MODAL,
-   };
-
-private:
-   type_t type;
-   placement3d place;
-
-public:
-   StageInterface(type_t type)
-      : type(type)
-      , place(0, 0, 0)
-   {}
-   virtual ~StageInterface() {}
-
-   type_t get_type() { return type; }
-   placement3d &get_place() { return place; }
-   void set_place(placement3d place) { this->place = place; }
-   bool infer_is_modal()
-   {
-      switch (type)
-      {
-      case ONE_LINE_INPUT_BOX:
-      case FILE_NAVIGATOR:
-      case KEYBOARD_INPUTS_MODAL:
-         return true;
-      default:
-        return false;
-      }
-   }
-
-   virtual void render(ALLEGRO_DISPLAY *display, ALLEGRO_FONT *font, int cell_width, int cell_height) = 0;
-   virtual void process_local_event(std::string event_name, ActionData action_data=ActionData()) = 0;
-   virtual void process_event(ALLEGRO_EVENT &event) = 0;
-
-   // actions
-
-   virtual bool save_file() = 0;
-};
-
+#include <Hexagon/StageInterface.hpp>
 
 
 class KeyboardInputsModal : public StageInterface
@@ -1144,124 +420,20 @@ public:
 
 
 
-class EventControllerInterface
-{
-public:
-   EventControllerInterface();
-   virtual ~EventControllerInterface();
-
-   virtual void process_local_event(std::string event_name, ActionData action_data1=ActionData()) = 0;
-   virtual void process_event(ALLEGRO_EVENT &event) = 0;
-};
+#include <Hexagon/EventControllerInterface.hpp>
 
 
 
-EventControllerInterface::EventControllerInterface() {}
-EventControllerInterface::~EventControllerInterface() {}
-
-
-
-class Stage;
-
-class StageRenderer
-{
-private:
-   ALLEGRO_DISPLAY *display;
-   RenderCache render_cache;
-   Stage *stage;
-   int cell_width;
-   int cell_height;
-   ALLEGRO_FONT *font;
-
-   void render_raw();
-   void draw_selections(int cell_width, int cell_height);
-
-public:
-   StageRenderer(Stage *stage, ALLEGRO_FONT *font, ALLEGRO_DISPLAY *display, int cell_width, int cell_height);
-   ~StageRenderer();
-
-   void render();
-};
-
-
-
-class StageEventController : public EventControllerInterface
-{
-private:
-   Stage *stage;
-
-   ActionQueueRecording last_performed_action_queue_recording;
-   bool last_performed_action_queue_is_recording;
-
-public:
-   StageEventController(Stage *stage);
-   virtual ~StageEventController();
-
-   bool clear_last_performed_action_queue_recording();
-   bool start_recording_last_performed_action_queue_recording();
-   bool stop_recording_last_performed_action_queue_recording();
-   bool is_last_performed_action_queue_recording();
-   bool play_last_performed_action_queue_recording();
-
-   static const std::string MOVE_CURSOR_UP;
-   static const std::string MOVE_CURSOR_DOWN;
-   static const std::string MOVE_CURSOR_LEFT;
-   static const std::string MOVE_CURSOR_RIGHT;
-   static const std::string MOVE_CURSOR_TO_TOP_OF_SCREEN;
-   static const std::string MOVE_CURSOR_TO_MIDDLE_OF_SCREEN;
-   static const std::string MOVE_CURSOR_TO_BOTTOM_OF_SCREEN;
-   static const std::string MOVE_CURSOR_JUMP_TO_NEXT_WORD;
-   static const std::string MOVE_CURSOR_JUMP_TO_NEXT_BIG_WORD;
-   static const std::string MOVE_CURSOR_JUMP_TO_PREVIOUS_WORD;
-   static const std::string JUMP_CURSOR_TO_END_OF_NEXT_WORD;
-   static const std::string JUMP_CURSOR_TO_END_OF_NEXT_BIG_WORD;
-   static const std::string JUMP_TO_NEXT_CODE_POINT;
-   static const std::string JUMP_TO_PREVIOUS_CODE_POINT;
-   static const std::string DELETE_CHARACTER;
-   static const std::string INSERT_STRING;
-   static const std::string SET_INSERT_MODE;
-   static const std::string SET_EDIT_MODE;
-   static const std::string JOIN_LINES;
-   static const std::string SPLIT_LINES;
-   static const std::string MOVE_CURSOR_TO_START_OF_LINE;
-   static const std::string MOVE_CURSOR_TO_END_OF_LINE;
-   static const std::string SAVE_FILE;
-   static const std::string MOVE_STAGE_UP;
-   static const std::string MOVE_STAGE_DOWN;
-   static const std::string SCALE_STAGE_UP;
-   static const std::string SCALE_STAGE_DOWN;
-   static const std::string JUMP_FIRST_LINE_NUM_DOWN;
-   static const std::string JUMP_FIRST_LINE_NUM_UP;
-   static const std::string JUMP_FIRST_LINE_NUM_DOWN_WHOLE_SCREEN;
-   static const std::string JUMP_FIRST_LINE_NUM_UP_WHOLE_SCREEN;
-   static const std::string STEP_FIRST_LINE_NUM_DOWN;
-   static const std::string STEP_FIRST_LINE_NUM_UP;
-   static const std::string OFFSET_CURSOR_POSITION_Y_DOWN;
-   static const std::string OFFSET_CURSOR_POSITION_Y_UP;
-   static const std::string OFFSET_CURSOR_POSITION_Y_DOWN_WHOLE_SCREEN;
-   static const std::string OFFSET_CURSOR_POSITION_Y_UP_WHOLE_SCREEN;
-   static const std::string REFRESH_GIT_MODIFIED_LINE_NUMBERS;
-   //static const std::string TOGGLE_SHOWING_CODE_MESSAGE_POINTS;
-   static const std::string REFRESH_REGEX_MESSAGE_POINTS;
-   static const std::string OFFSET_FIRST_LINE_TO_VERTICALLY_CENTER_CURSOR;
-   static const std::string CREATE_VISUAL_SELECTION_AT_CURRENT_CURSOR_LOCATION;
-   static const std::string DESTROY_CURRENT_VISUAL_SELECTION;
-   static const std::string TOGGLE_CURRENTLY_GRABBING_VISUAL_SELECTION;
-   static const std::string YANK_SELECTED_TEXT_TO_CLIPBOARD;
-   static const std::string PASTE_SELECTED_TEXT_FROM_CLIPBOARD;
-   static const std::string CLEAR_LAST_PERFORMED_ACTION_QUEUE_RECORDING;
-   static const std::string START_RECORDING_LAST_PERFORMED_ACTION_QUEUE_RECORDING;
-   static const std::string STOP_RECORDING_LAST_PERFORMED_ACTION_QUEUE_RECORDING;
-   static const std::string PLAY_LAST_PERFORMED_ACTION_QUEUE_RECORDING;
-
-   void process_local_event(std::string event_name, ActionData action_data1=ActionData()) override;
-   void process_event(ALLEGRO_EVENT &event) override;
-};
+#include <Hexagon/StageRenderer.hpp>
 
 
 
 
+#include <Hexagon/StageEventController.hpp>
 
+
+
+/* begin
 class Stage : public StageInterface
 {
 public:
@@ -1894,444 +1066,13 @@ public:
       stage_event_controller.process_event(event);
    }
 };
+*/
 
 
+#include <Hexagon/Stage.hpp>
 
 
-StageEventController::StageEventController(Stage *stage)
-   : EventControllerInterface()
-   , stage(stage)
-   , last_performed_action_queue_recording("last-performed-action-queue-recording")
-   , last_performed_action_queue_is_recording(false)
-{}
 
-
-
-StageEventController::~StageEventController()
-{}
-
-
-
-bool StageEventController::clear_last_performed_action_queue_recording()
-{
-   last_performed_action_queue_recording.clear_actions();
-   return true;
-}
-
-
-
-bool StageEventController::start_recording_last_performed_action_queue_recording()
-{
-   last_performed_action_queue_is_recording = true;
-   return true;
-}
-
-
-
-bool StageEventController::stop_recording_last_performed_action_queue_recording()
-{
-   last_performed_action_queue_is_recording = false;
-   return true;
-}
-
-
-
-bool StageEventController::is_last_performed_action_queue_recording()
-{
-   return last_performed_action_queue_is_recording;
-}
-
-
-
-bool StageEventController::play_last_performed_action_queue_recording()
-{
-   bool previous_last_performed_action_queue_is_recording = last_performed_action_queue_is_recording;
-   last_performed_action_queue_is_recording = false;
-
-   ActionQueueRecording recording = last_performed_action_queue_recording;
-
-   for (int i=0; i<recording.infer_num_actions(); i++)
-   {
-      Action action = recording.get_action_at(i);
-      process_local_event(action.get_name(), action.get_data1());
-   }
-
-   last_performed_action_queue_is_recording = previous_last_performed_action_queue_is_recording;
-
-   return true;
-}
-
-
-
-void StageEventController::process_local_event(std::string event_name, ActionData action_data1)
-{
-   if (!stage) throw std::runtime_error("[StageEventController::process_local_event()] error: stage cannot be a nullptr");
-
-   std::cout << "StageEventController::" << event_name << std::endl;
-
-   try {
-      if (event_name == MOVE_CURSOR_UP) stage->move_cursor_up();
-      else if (event_name == MOVE_CURSOR_DOWN) stage->move_cursor_down();
-      else if (event_name == MOVE_CURSOR_LEFT) stage->move_cursor_left();
-      else if (event_name == MOVE_CURSOR_RIGHT) stage->move_cursor_right();
-      else if (event_name == MOVE_CURSOR_TO_TOP_OF_SCREEN) stage->move_cursor_to_top_of_screen();
-      else if (event_name == MOVE_CURSOR_TO_MIDDLE_OF_SCREEN) stage->move_cursor_to_middle_of_screen();
-      else if (event_name == MOVE_CURSOR_TO_BOTTOM_OF_SCREEN) stage->move_cursor_to_bottom_of_screen();
-      else if (event_name == MOVE_CURSOR_JUMP_TO_NEXT_WORD) stage->move_cursor_jump_to_next_word();
-      else if (event_name == MOVE_CURSOR_JUMP_TO_NEXT_BIG_WORD) stage->move_cursor_jump_to_next_big_word();
-      else if (event_name == MOVE_CURSOR_JUMP_TO_PREVIOUS_WORD) stage->move_cursor_jump_to_previous_word();
-      else if (event_name == JUMP_CURSOR_TO_END_OF_NEXT_WORD) stage->jump_cursor_to_end_of_next_word();
-      else if (event_name == JUMP_CURSOR_TO_END_OF_NEXT_BIG_WORD) stage->jump_cursor_to_end_of_next_big_word();
-      else if (event_name == JUMP_TO_NEXT_CODE_POINT) stage->jump_to_next_code_point();
-      else if (event_name == JUMP_TO_PREVIOUS_CODE_POINT) stage->jump_to_previous_code_point();
-      else if (event_name == DELETE_CHARACTER) stage->delete_character();
-      else if (event_name == INSERT_STRING) stage->insert_string(action_data1.get_string());
-      else if (event_name == SET_INSERT_MODE) stage->set_insert_mode();
-      else if (event_name == SET_EDIT_MODE) stage->set_edit_mode();
-      else if (event_name == JOIN_LINES) stage->join_lines();
-      else if (event_name == SPLIT_LINES) stage->split_lines();
-      else if (event_name == MOVE_CURSOR_TO_START_OF_LINE) stage->move_cursor_to_start_of_line();
-      else if (event_name == MOVE_CURSOR_TO_END_OF_LINE) stage->move_cursor_to_end_of_line();
-      else if (event_name == SAVE_FILE) stage->save_file();
-      else if (event_name == MOVE_STAGE_UP) stage->move_stage_up();
-      else if (event_name == MOVE_STAGE_DOWN) stage->move_stage_down();
-      else if (event_name == SCALE_STAGE_UP) stage->scale_stage_delta(0.1);
-      else if (event_name == SCALE_STAGE_DOWN) stage->scale_stage_delta(-0.1);
-      else if (event_name == JUMP_FIRST_LINE_NUM_DOWN_WHOLE_SCREEN) stage->offset_first_line_number(stage->infer_num_lines_to_draw());
-      else if (event_name == JUMP_FIRST_LINE_NUM_UP_WHOLE_SCREEN) stage->offset_first_line_number(-stage->infer_num_lines_to_draw());
-      else if (event_name == JUMP_FIRST_LINE_NUM_DOWN) stage->offset_first_line_number(stage->infer_num_lines_to_draw()/2);
-      else if (event_name == JUMP_FIRST_LINE_NUM_UP) stage->offset_first_line_number(-stage->infer_num_lines_to_draw()/2);
-      else if (event_name == STEP_FIRST_LINE_NUM_DOWN) stage->offset_first_line_number(1);
-      else if (event_name == STEP_FIRST_LINE_NUM_UP) stage->offset_first_line_number(-1);
-      else if (event_name == OFFSET_CURSOR_POSITION_Y_DOWN) stage->offset_cursor_position_y(stage->infer_num_lines_to_draw()/2);
-      else if (event_name == OFFSET_CURSOR_POSITION_Y_UP) stage->offset_cursor_position_y(-stage->infer_num_lines_to_draw()/2);
-      else if (event_name == OFFSET_CURSOR_POSITION_Y_DOWN_WHOLE_SCREEN) stage->offset_cursor_position_y(stage->infer_num_lines_to_draw());
-      else if (event_name == OFFSET_CURSOR_POSITION_Y_UP_WHOLE_SCREEN) stage->offset_cursor_position_y(-stage->infer_num_lines_to_draw());
-      else if (event_name == REFRESH_GIT_MODIFIED_LINE_NUMBERS) stage->refresh_git_modified_line_numbers();
-      //else if (event_name == TOGGLE_SHOWING_CODE_MESSAGE_POINTS) toggle_showing_code_message_points();
-      else if (event_name == REFRESH_REGEX_MESSAGE_POINTS) stage->refresh_regex_message_points();
-      else if (event_name == OFFSET_FIRST_LINE_TO_VERTICALLY_CENTER_CURSOR) stage->offset_first_line_to_vertically_center_cursor();
-      else if (event_name == CREATE_VISUAL_SELECTION_AT_CURRENT_CURSOR_LOCATION) stage->create_visual_selection_at_current_cursor_location();
-      else if (event_name == DESTROY_CURRENT_VISUAL_SELECTION) stage->destroy_current_visual_selection();
-      else if (event_name == TOGGLE_CURRENTLY_GRABBING_VISUAL_SELECTION) stage->toggle_currently_grabbing_visual_selection();
-      else if (event_name == YANK_SELECTED_TEXT_TO_CLIPBOARD) stage->yank_selected_text_to_clipboard();
-      else if (event_name == PASTE_SELECTED_TEXT_FROM_CLIPBOARD) stage->paste_selected_text_from_clipboard();
-      else if (event_name == CLEAR_LAST_PERFORMED_ACTION_QUEUE_RECORDING) clear_last_performed_action_queue_recording();
-      else if (event_name == START_RECORDING_LAST_PERFORMED_ACTION_QUEUE_RECORDING) start_recording_last_performed_action_queue_recording();
-      else if (event_name == STOP_RECORDING_LAST_PERFORMED_ACTION_QUEUE_RECORDING) stop_recording_last_performed_action_queue_recording();
-      else if (event_name == PLAY_LAST_PERFORMED_ACTION_QUEUE_RECORDING) play_last_performed_action_queue_recording();
-   }
-
-   catch (const std::exception &e)
-   {
-      std::cout << "\033[0;33m💥 cannot execute \"" << event_name << "\": \033[0;31m" << e.what() << "\033[0;33m\033[0m" << std::endl;
-   }
-
-   if (last_performed_action_queue_is_recording && event_name != START_RECORDING_LAST_PERFORMED_ACTION_QUEUE_RECORDING)
-   {
-      std::cout << "   \033[0;30mStage:: recording " << event_name << " event to last_performed_action_queue_recording queue\033[0m" << std::endl;
-      last_performed_action_queue_recording.append_action(Action(event_name, action_data1));
-   }
-}
-
-void StageEventController::process_event(ALLEGRO_EVENT &event)
-{
-   //std::map<std::tuple<int, bool, bool, bool>, std::vector<std::string>> mapping;
-   //bool set_mapping(int al_keycode, bool shift, bool ctrl, bool alt, std::vector<std::string> comand_identifier);
-   KeyboardCommandMapper edit_mode__keyboard_command_mapper;
-   edit_mode__keyboard_command_mapper.set_mapping(ALLEGRO_KEY_0, false, false, false, { StageEventController::MOVE_CURSOR_TO_START_OF_LINE });
-   edit_mode__keyboard_command_mapper.set_mapping(ALLEGRO_KEY_I, true,  false, false, { StageEventController::CLEAR_LAST_PERFORMED_ACTION_QUEUE_RECORDING, StageEventController::START_RECORDING_LAST_PERFORMED_ACTION_QUEUE_RECORDING, StageEventController::MOVE_CURSOR_TO_START_OF_LINE, StageEventController::MOVE_CURSOR_JUMP_TO_NEXT_WORD, StageEventController::SET_INSERT_MODE, });
-   edit_mode__keyboard_command_mapper.set_mapping(ALLEGRO_KEY_A, true,  false, false, { StageEventController::CLEAR_LAST_PERFORMED_ACTION_QUEUE_RECORDING, StageEventController::START_RECORDING_LAST_PERFORMED_ACTION_QUEUE_RECORDING, StageEventController::MOVE_CURSOR_TO_END_OF_LINE, StageEventController::SET_INSERT_MODE, });
-   edit_mode__keyboard_command_mapper.set_mapping(ALLEGRO_KEY_A, false,  false, false, { StageEventController::CLEAR_LAST_PERFORMED_ACTION_QUEUE_RECORDING, StageEventController::START_RECORDING_LAST_PERFORMED_ACTION_QUEUE_RECORDING, StageEventController::MOVE_CURSOR_RIGHT, StageEventController::SET_INSERT_MODE, });
-   edit_mode__keyboard_command_mapper.set_mapping(ALLEGRO_KEY_J, false, false, false, { StageEventController::MOVE_CURSOR_DOWN });
-   edit_mode__keyboard_command_mapper.set_mapping(ALLEGRO_KEY_J, true,  false, false, { StageEventController::JOIN_LINES });
-   edit_mode__keyboard_command_mapper.set_mapping(ALLEGRO_KEY_K, false, false, false, { StageEventController::MOVE_CURSOR_UP });
-   edit_mode__keyboard_command_mapper.set_mapping(ALLEGRO_KEY_H, false, false, false, { StageEventController::MOVE_CURSOR_LEFT });
-   edit_mode__keyboard_command_mapper.set_mapping(ALLEGRO_KEY_L, false, false, false, { StageEventController::MOVE_CURSOR_RIGHT });
-   edit_mode__keyboard_command_mapper.set_mapping(ALLEGRO_KEY_H, true, false, false, { StageEventController::MOVE_CURSOR_TO_TOP_OF_SCREEN });
-   edit_mode__keyboard_command_mapper.set_mapping(ALLEGRO_KEY_M, true, false, false, { StageEventController::MOVE_CURSOR_TO_MIDDLE_OF_SCREEN });
-   edit_mode__keyboard_command_mapper.set_mapping(ALLEGRO_KEY_L, true, false, false, { StageEventController::MOVE_CURSOR_TO_BOTTOM_OF_SCREEN });
-   edit_mode__keyboard_command_mapper.set_mapping(ALLEGRO_KEY_W, false, false, false, { StageEventController::MOVE_CURSOR_JUMP_TO_NEXT_WORD });
-   edit_mode__keyboard_command_mapper.set_mapping(ALLEGRO_KEY_W, true, false, false, { StageEventController::MOVE_CURSOR_JUMP_TO_NEXT_BIG_WORD });
-   edit_mode__keyboard_command_mapper.set_mapping(ALLEGRO_KEY_E, false, false, false, { StageEventController::JUMP_CURSOR_TO_END_OF_NEXT_WORD });
-   edit_mode__keyboard_command_mapper.set_mapping(ALLEGRO_KEY_E, true, false, false, { StageEventController::JUMP_CURSOR_TO_END_OF_NEXT_BIG_WORD });
-   edit_mode__keyboard_command_mapper.set_mapping(ALLEGRO_KEY_B, false, false, false, { StageEventController::MOVE_CURSOR_JUMP_TO_PREVIOUS_WORD });
-   edit_mode__keyboard_command_mapper.set_mapping(ALLEGRO_KEY_N, false, false, false, { StageEventController::JUMP_TO_NEXT_CODE_POINT, StageEventController::OFFSET_FIRST_LINE_TO_VERTICALLY_CENTER_CURSOR });
-   edit_mode__keyboard_command_mapper.set_mapping(ALLEGRO_KEY_N, true, false, false, { StageEventController::JUMP_TO_PREVIOUS_CODE_POINT, StageEventController::OFFSET_FIRST_LINE_TO_VERTICALLY_CENTER_CURSOR });
-   edit_mode__keyboard_command_mapper.set_mapping(ALLEGRO_KEY_X, false, false, false, { StageEventController::DELETE_CHARACTER });
-   edit_mode__keyboard_command_mapper.set_mapping(ALLEGRO_KEY_I, false, false, false, { StageEventController::CLEAR_LAST_PERFORMED_ACTION_QUEUE_RECORDING, StageEventController::START_RECORDING_LAST_PERFORMED_ACTION_QUEUE_RECORDING, StageEventController::SET_INSERT_MODE, });
-   edit_mode__keyboard_command_mapper.set_mapping(ALLEGRO_KEY_0, false, false, false, { StageEventController::MOVE_CURSOR_TO_START_OF_LINE });
-   edit_mode__keyboard_command_mapper.set_mapping(ALLEGRO_KEY_S, false, false, false, { StageEventController::CLEAR_LAST_PERFORMED_ACTION_QUEUE_RECORDING, StageEventController::START_RECORDING_LAST_PERFORMED_ACTION_QUEUE_RECORDING, StageEventController::DELETE_CHARACTER, StageEventController::SET_INSERT_MODE, });
-   edit_mode__keyboard_command_mapper.set_mapping(ALLEGRO_KEY_S, false, true, false, { StageEventController::SAVE_FILE });
-   edit_mode__keyboard_command_mapper.set_mapping(ALLEGRO_KEY_B, false, true, false, { StageEventController::JUMP_FIRST_LINE_NUM_UP_WHOLE_SCREEN, StageEventController::OFFSET_CURSOR_POSITION_Y_UP_WHOLE_SCREEN });
-   edit_mode__keyboard_command_mapper.set_mapping(ALLEGRO_KEY_F, false, true, false, { StageEventController::JUMP_FIRST_LINE_NUM_DOWN_WHOLE_SCREEN, StageEventController::OFFSET_CURSOR_POSITION_Y_DOWN_WHOLE_SCREEN });
-   edit_mode__keyboard_command_mapper.set_mapping(ALLEGRO_KEY_U, false, true, false, { StageEventController::JUMP_FIRST_LINE_NUM_UP, StageEventController::OFFSET_CURSOR_POSITION_Y_UP });
-   edit_mode__keyboard_command_mapper.set_mapping(ALLEGRO_KEY_D, false, true, false, { StageEventController::JUMP_FIRST_LINE_NUM_DOWN, StageEventController::OFFSET_CURSOR_POSITION_Y_DOWN });
-   edit_mode__keyboard_command_mapper.set_mapping(ALLEGRO_KEY_Y, false, true, false, { StageEventController::STEP_FIRST_LINE_NUM_UP });
-   edit_mode__keyboard_command_mapper.set_mapping(ALLEGRO_KEY_E, false, true, false, { StageEventController::STEP_FIRST_LINE_NUM_DOWN });
-   edit_mode__keyboard_command_mapper.set_mapping(ALLEGRO_KEY_EQUALS, false, true, false, { StageEventController::SCALE_STAGE_UP });
-   edit_mode__keyboard_command_mapper.set_mapping(ALLEGRO_KEY_MINUS, false, true, false, { StageEventController::SCALE_STAGE_DOWN });
-   edit_mode__keyboard_command_mapper.set_mapping(ALLEGRO_KEY_O, false, false, false, { StageEventController::CLEAR_LAST_PERFORMED_ACTION_QUEUE_RECORDING, StageEventController::START_RECORDING_LAST_PERFORMED_ACTION_QUEUE_RECORDING, StageEventController::MOVE_CURSOR_TO_END_OF_LINE, StageEventController::SPLIT_LINES, StageEventController::MOVE_CURSOR_DOWN, StageEventController::MOVE_CURSOR_TO_START_OF_LINE, StageEventController::SET_INSERT_MODE, });
-   edit_mode__keyboard_command_mapper.set_mapping(ALLEGRO_KEY_O, true, false, false, { StageEventController::CLEAR_LAST_PERFORMED_ACTION_QUEUE_RECORDING, StageEventController::START_RECORDING_LAST_PERFORMED_ACTION_QUEUE_RECORDING, StageEventController::MOVE_CURSOR_UP, StageEventController::MOVE_CURSOR_TO_END_OF_LINE, StageEventController::SPLIT_LINES, StageEventController::MOVE_CURSOR_DOWN, StageEventController::MOVE_CURSOR_TO_START_OF_LINE, StageEventController::SET_INSERT_MODE, });
-   edit_mode__keyboard_command_mapper.set_mapping(ALLEGRO_KEY_G, false, false, true, { StageEventController::SAVE_FILE, StageEventController::REFRESH_GIT_MODIFIED_LINE_NUMBERS });
-   //edit_mode__keyboard_command_mapper.set_mapping(ALLEGRO_KEY_TAB, false, false, false, { StageEventController::TOGGLE_SHOWING_CODE_MESSAGE_POINTS });
-   edit_mode__keyboard_command_mapper.set_mapping(ALLEGRO_KEY_SLASH, false, false, false, { StageEventController::REFRESH_REGEX_MESSAGE_POINTS });
-   edit_mode__keyboard_command_mapper.set_mapping(ALLEGRO_KEY_Z, false, false, false, { StageEventController::OFFSET_FIRST_LINE_TO_VERTICALLY_CENTER_CURSOR });
-   edit_mode__keyboard_command_mapper.set_mapping(ALLEGRO_KEY_V, false, false, false, { StageEventController::TOGGLE_CURRENTLY_GRABBING_VISUAL_SELECTION });
-   edit_mode__keyboard_command_mapper.set_mapping(ALLEGRO_KEY_Y, false, false, false, { StageEventController::YANK_SELECTED_TEXT_TO_CLIPBOARD });
-   edit_mode__keyboard_command_mapper.set_mapping(ALLEGRO_KEY_P, true, false, false, { StageEventController::PASTE_SELECTED_TEXT_FROM_CLIPBOARD });
-   edit_mode__keyboard_command_mapper.set_mapping(ALLEGRO_KEY_FULLSTOP, false, false, false, { StageEventController::PLAY_LAST_PERFORMED_ACTION_QUEUE_RECORDING });
-
-
-   //std::map<std::tuple<int, bool, bool, bool>, std::vector<std::string>> mapping;
-   //bool set_mapping(int al_keycode, bool shift, bool ctrl, bool alt, std::vector<std::string> comand_identifier);
-   KeyboardCommandMapper insert_mode__keyboard_command_mapper;
-   insert_mode__keyboard_command_mapper.set_mapping(ALLEGRO_KEY_ESCAPE, false, false, false, { StageEventController::SET_EDIT_MODE, StageEventController::STOP_RECORDING_LAST_PERFORMED_ACTION_QUEUE_RECORDING });
-   insert_mode__keyboard_command_mapper.set_mapping(ALLEGRO_KEY_OPENBRACE, false, true, false, { StageEventController::SET_EDIT_MODE, StageEventController::STOP_RECORDING_LAST_PERFORMED_ACTION_QUEUE_RECORDING });
-   insert_mode__keyboard_command_mapper.set_mapping(ALLEGRO_KEY_BACKSPACE, false, false, false, { StageEventController::MOVE_CURSOR_LEFT, StageEventController::DELETE_CHARACTER });
-   insert_mode__keyboard_command_mapper.set_mapping(ALLEGRO_KEY_ENTER, false, false, false, { StageEventController::SPLIT_LINES, StageEventController::MOVE_CURSOR_DOWN, StageEventController::MOVE_CURSOR_TO_START_OF_LINE });
-   insert_mode__keyboard_command_mapper.set_mapping(ALLEGRO_KEY_S, false, true, false, { StageEventController::SAVE_FILE });
-
-
-   switch(stage->get_mode())
-   {
-   case Stage::EDIT:
-      switch(event.type)
-      {
-      case ALLEGRO_EVENT_KEY_DOWN:
-         break;
-      case ALLEGRO_EVENT_KEY_UP:
-         break;
-      case ALLEGRO_EVENT_KEY_CHAR:
-         bool shift = event.keyboard.modifiers & ALLEGRO_KEYMOD_SHIFT;
-         bool ctrl = event.keyboard.modifiers & ALLEGRO_KEYMOD_CTRL || event.keyboard.modifiers & ALLEGRO_KEYMOD_COMMAND;
-         bool alt = event.keyboard.modifiers & ALLEGRO_KEYMOD_ALT;
-         std::vector<std::string> mapped_events = edit_mode__keyboard_command_mapper.get_mapping(event.keyboard.keycode, shift, ctrl, alt);
-         for (auto &mapped_event : mapped_events) process_local_event(mapped_event);
-         break;
-      }
-      break;
-   case Stage::INSERT:
-      switch(event.type)
-      {
-      case ALLEGRO_EVENT_KEY_DOWN:
-         break;
-      case ALLEGRO_EVENT_KEY_UP:
-         break;
-      case ALLEGRO_EVENT_KEY_CHAR:
-         bool shift = event.keyboard.modifiers & ALLEGRO_KEYMOD_SHIFT;
-         bool ctrl = event.keyboard.modifiers & ALLEGRO_KEYMOD_CTRL || event.keyboard.modifiers & ALLEGRO_KEYMOD_COMMAND;
-         bool alt = event.keyboard.modifiers & ALLEGRO_KEYMOD_ALT;
-         std::vector<std::string> mapped_events = insert_mode__keyboard_command_mapper.get_mapping(event.keyboard.keycode, shift, ctrl, alt);
-         for (auto &mapped_event : mapped_events) process_local_event(mapped_event);
-         if (mapped_events.empty())
-         {
-            char character = (char)(event.keyboard.unichar);
-            std::string string = " ";
-            string[0] = character;
-            process_local_event(INSERT_STRING, ActionData(string));
-            process_local_event(MOVE_CURSOR_RIGHT);
-         }
-         break;
-      }
-      break;
-   }
-}
-
-
-
-std::string const StageEventController::MOVE_CURSOR_UP = "MOVE_CURSOR_UP";
-std::string const StageEventController::MOVE_CURSOR_DOWN = "MOVE_CURSOR_DOWN";
-std::string const StageEventController::MOVE_CURSOR_LEFT = "MOVE_CURSOR_LEFT";
-std::string const StageEventController::MOVE_CURSOR_RIGHT = "MOVE_CURSOR_RIGHT";
-std::string const StageEventController::MOVE_CURSOR_TO_TOP_OF_SCREEN = "MOVE_CURSOR_TO_TOP_OF_SCREEN";
-std::string const StageEventController::MOVE_CURSOR_TO_MIDDLE_OF_SCREEN = "MOVE_CURSOR_TO_MIDDLE_OF_SCREEN";
-std::string const StageEventController::MOVE_CURSOR_TO_BOTTOM_OF_SCREEN = "MOVE_CURSOR_TO_BOTTOM_OF_SCREEN";
-std::string const StageEventController::MOVE_CURSOR_JUMP_TO_NEXT_WORD = "MOVE_CURSOR_JUMP_TO_NEXT_WORD";
-std::string const StageEventController::MOVE_CURSOR_JUMP_TO_NEXT_BIG_WORD = "MOVE_CURSOR_JUMP_TO_NEXT_BIG_WORD";
-std::string const StageEventController::MOVE_CURSOR_JUMP_TO_PREVIOUS_WORD = "MOVE_CURSOR_JUMP_TO_PREVIOUS_WORD";
-std::string const StageEventController::JUMP_CURSOR_TO_END_OF_NEXT_WORD = "JUMP_CURSOR_TO_END_OF_NEXT_WORD";
-std::string const StageEventController::JUMP_CURSOR_TO_END_OF_NEXT_BIG_WORD = "JUMP_CURSOR_TO_END_OF_NEXT_BIG_WORD";
-std::string const StageEventController::JUMP_TO_NEXT_CODE_POINT = "JUMP_TO_NEXT_CODE_POINT";
-std::string const StageEventController::JUMP_TO_PREVIOUS_CODE_POINT = "JUMP_TO_PREVIOUS_CODE_POINT";
-std::string const StageEventController::DELETE_CHARACTER = "DELETE_CHARACTER";
-std::string const StageEventController::INSERT_STRING = "INSERT_STRING";
-std::string const StageEventController::SET_INSERT_MODE = "SET_INSERT_MODE";
-std::string const StageEventController::SET_EDIT_MODE = "SET_EDIT_MODE";
-std::string const StageEventController::JOIN_LINES = "JOIN_LINES";
-std::string const StageEventController::SPLIT_LINES = "SPLIT_LINES";
-std::string const StageEventController::MOVE_CURSOR_TO_START_OF_LINE = "MOVE_CURSOR_TO_START_OF_LINE";
-std::string const StageEventController::MOVE_CURSOR_TO_END_OF_LINE = "MOVE_CURSOR_TO_END_OF_LINE";
-std::string const StageEventController::SAVE_FILE = "SAVE_FILE";
-std::string const StageEventController::MOVE_STAGE_UP = "MOVE_STAGE_UP";
-std::string const StageEventController::MOVE_STAGE_DOWN = "MOVE_STAGE_DOWN";
-std::string const StageEventController::JUMP_FIRST_LINE_NUM_UP = "JUMP_FIRST_LINE_NUM_UP";
-std::string const StageEventController::JUMP_FIRST_LINE_NUM_DOWN = "JUMP_FIRST_LINE_NUM_DOWN";
-std::string const StageEventController::JUMP_FIRST_LINE_NUM_UP_WHOLE_SCREEN = "JUMP_FIRST_LINE_NUM_UP_WHOLE_SCREEN";
-std::string const StageEventController::JUMP_FIRST_LINE_NUM_DOWN_WHOLE_SCREEN = "JUMP_FIRST_LINE_NUM_DOWN_WHOLE_SCREEN";
-std::string const StageEventController::STEP_FIRST_LINE_NUM_UP = "STEP_FIRST_LINE_NUM_UP";
-std::string const StageEventController::STEP_FIRST_LINE_NUM_DOWN = "STEP_FIRST_LINE_NUM_DOWN";
-std::string const StageEventController::OFFSET_CURSOR_POSITION_Y_UP = "OFFSET_CURSOR_POSITION_Y_UP";
-std::string const StageEventController::OFFSET_CURSOR_POSITION_Y_DOWN = "OFFSET_CURSOR_POSITION_Y_DOWN";
-std::string const StageEventController::OFFSET_CURSOR_POSITION_Y_UP_WHOLE_SCREEN = "OFFSET_CURSOR_POSITION_Y_UP_WHOLE_SCREEN";
-std::string const StageEventController::OFFSET_CURSOR_POSITION_Y_DOWN_WHOLE_SCREEN = "OFFSET_CURSOR_POSITION_Y_DOWN_WHOLE_SCREEN";
-std::string const StageEventController::SCALE_STAGE_UP = "SCALE_STAGE_UP";
-std::string const StageEventController::SCALE_STAGE_DOWN = "SCALE_STAGE_DOWN";
-std::string const StageEventController::REFRESH_GIT_MODIFIED_LINE_NUMBERS = "REFRESH_GIT_MODIFIED_LINE_NUMBERS";
-//std::string const StageEventController::TOGGLE_SHOWING_CODE_MESSAGE_POINTS = "TOGGLE_SHOWING_CODE_MESSAGE_POINTS";
-std::string const StageEventController::REFRESH_REGEX_MESSAGE_POINTS = "REFRESH_REGEX_MESSAGE_POINTS";
-std::string const StageEventController::OFFSET_FIRST_LINE_TO_VERTICALLY_CENTER_CURSOR = "OFFSET_FIRST_LINE_TO_VERTICALLY_CENTER_CURSOR";
-std::string const StageEventController::CREATE_VISUAL_SELECTION_AT_CURRENT_CURSOR_LOCATION = "CREATE_VISUAL_SELECTION_AT_CURRENT_CURSOR_LOCATION";
-std::string const StageEventController::DESTROY_CURRENT_VISUAL_SELECTION = "DESTROY_CURRENT_VISUAL_SELECTION";
-std::string const StageEventController::TOGGLE_CURRENTLY_GRABBING_VISUAL_SELECTION = "TOGGLE_CURRENTLY_GRABBING_VISUAL_SELECTION";
-std::string const StageEventController::YANK_SELECTED_TEXT_TO_CLIPBOARD = "YANK_SELECTED_TEXT_TO_CLIPBOARD";
-std::string const StageEventController::PASTE_SELECTED_TEXT_FROM_CLIPBOARD = "PASTE_SELECTED_TEXT_FROM_CLIPBOARD";
-std::string const StageEventController::CLEAR_LAST_PERFORMED_ACTION_QUEUE_RECORDING = "CLEAR_LAST_PERFORMED_ACTION_QUEUE_RECORDING";
-std::string const StageEventController::START_RECORDING_LAST_PERFORMED_ACTION_QUEUE_RECORDING = "START_RECORDING_LAST_PERFORMED_ACTION_QUEUE_RECORDING";
-std::string const StageEventController::STOP_RECORDING_LAST_PERFORMED_ACTION_QUEUE_RECORDING = "STOP_RECORDING_LAST_PERFORMED_ACTION_QUEUE_RECORDING";
-std::string const StageEventController::PLAY_LAST_PERFORMED_ACTION_QUEUE_RECORDING = "PLAY_LAST_PERFORMED_ACTION_QUEUE_RECORDING";
-
-
-
-void StageRenderer::draw_selections(int cell_width, int cell_height)
-{
-   for (auto &selection : stage->selections)
-   {
-      //std::cout << " drawing selection " << selection << std::endl;
-      CodeRangeRenderer(stage->get_lines_ref(), selection, stage->first_line_number, cell_width, cell_height).render();
-   }
-}
-
-
-
-void StageRenderer::render_raw()
-{
-      if (!stage) throw std::runtime_error("[StageRenderer] stage cannot be a nullptr");
-
-      //get_place().start_transform();
-      placement3d &place = stage->get_place();
-
-      //place.start_transform();
-
-      ALLEGRO_COLOR background_overlay_color = al_color_name("black");
-      float opacity = 0.8;
-      background_overlay_color.r *= opacity;
-      background_overlay_color.g *= opacity;
-      background_overlay_color.b *= opacity;
-      background_overlay_color.a *= opacity;
-
-      //al_draw_filled_rectangle(0, 0, al_get_display_width(display), al_get_display_height(display), background_overlay_color);
-
-      //get_place().start_transform();
-      al_draw_filled_rectangle(0, 0, place.size.x, place.size.y, background_overlay_color);
-
-      // draw a frame around the stage
-      //float padding = 30;
-      //float roundness = 20;
-      //float thickness = 20.0;
-      //al_draw_rounded_rectangle(-padding, -padding, al_get_display_width(display)+padding, al_get_display_height(display)+padding, roundness, roundness, al_color_name("deepskyblue"), thickness);
-
-      // render cursor
-      int &first_line_number = stage->first_line_number;
-      float _cursor_y = stage->cursor_y - stage->first_line_number;
-      int cursor_x = stage->cursor_x;
-      float _cell_width = cell_width;
-      float _cell_height = cell_height;
-      Stage::mode_t mode = stage->mode;
-
-      switch(mode)
-      {
-      case Stage::EDIT:
-         al_draw_filled_rectangle(cursor_x*_cell_width, _cursor_y*cell_height, cursor_x*_cell_width + _cell_width, _cursor_y*cell_height + cell_height, al_color_name("gray"));
-         break;
-      case Stage::INSERT:
-         al_draw_line(cursor_x*_cell_width, _cursor_y*cell_height, cursor_x*_cell_width, _cursor_y*cell_height + cell_height, al_color_name("gray"), 3);
-         break;
-      }
-
-      draw_selections(_cell_width, cell_height);
-
-      // render lines
-      int line_height = al_get_font_line_height(font);
-      const int line_count_render_limit = 120;
-      int lines_rendered_count = 0;
-      std::vector<int> &git_modified_line_numbers = stage->git_modified_line_numbers;
-      std::vector<CodeMessagePointsOverlay> &code_message_points_overlays = stage->code_message_points_overlays;
-
-
-      for (int line_number = first_line_number; line_number < (int)stage->lines.size(); line_number++)
-      {
-         bool line_exists_in_git_modified_line_numbers = std::find(git_modified_line_numbers.begin(), git_modified_line_numbers.end(), (line_number+1)) != git_modified_line_numbers.end();
-         if (line_exists_in_git_modified_line_numbers)
-         {
-            ALLEGRO_COLOR color = al_color_name("orange");
-            color.r *= 0.13;
-            color.g *= 0.13;
-            color.b *= 0.13;
-            color.a *= 0.13;
-            al_draw_filled_rectangle(0, line_height * (line_number - first_line_number), al_get_display_width(display)*2, line_height * (line_number - first_line_number + 1), color);
-         }
-
-         if (line_number >= 0) al_draw_text(font, al_color_name("white"), 0, (line_number-first_line_number)*cell_height, ALLEGRO_ALIGN_LEFT, stage->lines[line_number].c_str());
-         std::stringstream ss;
-         ss << (line_number+1);
-         ALLEGRO_COLOR text_color = al_color_name("darkolivegreen");
-         if (line_exists_in_git_modified_line_numbers) text_color = al_color_name("orange");
-         al_draw_text(font, text_color, -20, (line_number-first_line_number)*cell_height, ALLEGRO_ALIGN_RIGHT, ss.str().c_str());
-
-         lines_rendered_count++;
-         if (lines_rendered_count >= line_count_render_limit) break;
-      }
-
-      for (auto &code_message_points_overlay : code_message_points_overlays)
-      {
-         code_message_points_overlay.render(font, first_line_number, al_get_font_line_height(font), cursor_x, _cursor_y);
-      }
-      //for (auto &code_message_point : code_message_points)
-      //{
-         //CodeMessagePointRenderer code_message_point_renderer(code_message_point, font, first_line_number, al_get_font_line_height(font));
-         //code_message_point_renderer.render();
-      //}
-
-      //stage->get_place().restore_transform();
-
-      // render edit mode
-      ALLEGRO_COLOR color;
-
-      switch(mode)
-      {
-      case Stage::EDIT: color = (stage->currently_grabbing_visual_selection ? al_color_name("orange") : al_color_name("red")); break;
-      case Stage::INSERT: color = al_color_name("lime"); break;
-      default: break;
-      }
-
-      al_draw_text(font, color, al_get_display_width(display)/2, al_get_display_height(display)-al_get_font_line_height(font)*2, ALLEGRO_ALIGN_CENTER, stage->get_current_mode_string().c_str());
-      al_draw_text(font, al_color_name("gray"), al_get_display_width(display)/2, al_get_display_height(display)-al_get_font_line_height(font)*1, ALLEGRO_ALIGN_CENTER, stage->filename.c_str());
-}
-
-StageRenderer::StageRenderer(Stage *stage, ALLEGRO_FONT *font, ALLEGRO_DISPLAY *display, int cell_width, int cell_height)
-   : stage(stage)
-   , font(font)
-   , display(display)
-   , render_cache()
-   , cell_width(cell_width)
-   , cell_height(cell_height)
-{}
-
-StageRenderer::~StageRenderer() {}
-
-void StageRenderer::render()
-{
-   placement3d &place = stage->get_place();
-
-   render_cache.setup_surface(place.size.x, place.size.y);
-   render_raw();
-   render_cache.finish_surface();
-
-   place.start_transform();
-   render_cache.draw(0, 0);
-   place.restore_transform();
-}
