+++ src/Hexagon/EventControllerInterface.cpp
@@ -1,0 +1,17 @@
+
+
+
+#include <Hexagon/EventControllerInterface.hpp>
+
+
+
+EventControllerInterface::EventControllerInterface()
+{}
+
+
+
+EventControllerInterface::~EventControllerInterface()
+{}
+
+
+
