+++ src/Hexagon/shared_globals.cpp
@@ -1,0 +1,14 @@
+
+
+
+#include <Hexagon/shared_globals.hpp>
+
+
+
+std::string REGEX_TEMP_FILENAME = "regex.txt";
+std::string CLIPBOARD_TEMP_FILENAME = "clipboard.txt";
+std::string FILE_NAVIGATOR_SELECTION_FILENAME = "file_navigator_selection.txt";
+std::string FILE_NAVIGATOR_SELECTION_last_content = "";
+
+
+
