+++ src/Hexagon/util.cpp
@@ -1,0 +1,100 @@
+
+
+
+#include <Hexagon/util.hpp>
+
+
+#include <iostream>       // std::cout
+#include <string>         // std::string
+#include <cstddef>        // std::size_t
+#include <fstream>
+#include <sstream>
+
+
+std::string::size_type rfind_first_not_of(const std::string& s,
+        const std::string& delims, std::string::size_type pos)
+{
+    std::string::size_type p = pos;
+    while (delims.find(s[p]) != std::string::npos) {
+        if (p == 0) // not found
+            return std::string::npos;
+        --p;
+    }
+    return p;
+}
+
+std::string::size_type rfind_first_of(const std::string& s,
+        const std::string& delims, std::string::size_type pos)
+{
+    std::string::size_type p = pos;
+    while (delims.find(s[p]) == std::string::npos) {
+        if (p == 0) // not found
+            return std::string::npos;
+        --p;
+    }
+    return p;
+}
+
+int find_whitespace(std::string &string, int current_cursor_position)
+{
+  std::size_t position = string.find_first_of(" \n\r\t", current_cursor_position);
+  if (position != std::string::npos) return position;
+  return 0;
+}
+
+int find_non_whitespace(std::string &string, int current_cursor_position)
+{
+  std::size_t position = string.find_first_not_of(" \n\r\t", current_cursor_position);
+  if (position != std::string::npos) return position;
+  return 0;
+}
+
+int rfind_whitespace(std::string &string, int current_cursor_position)
+{
+  std::size_t position = rfind_first_of(string, " \n\r\t", current_cursor_position);
+  if (position != std::string::npos) return position;
+  return 0;
+}
+
+int rfind_non_whitespace(std::string &string, int current_cursor_position)
+{
+  std::size_t position = rfind_first_not_of(string, " \n\r\t", current_cursor_position);
+  if (position != std::string::npos) return position;
+  return 0;
+}
+
+
+bool read_file(std::vector<std::string> &lines, std::string filename)
+{
+   std::ifstream file(filename);
+
+   lines.clear();
+
+   if (file.is_open())
+   {
+      std::string line;
+      while (getline(file, line)) { lines.push_back(line); }
+      file.close();
+   }
+
+   return true;
+}
+
+bool save_file(std::vector<std::string> &lines, std::string filename)
+{
+   std::ofstream file(filename);
+
+   if (file.is_open())
+   {
+      for (auto &line : lines) file << line << '\n';
+      file.close();
+   }
+   else
+   {
+      std::stringstream error_message;
+      error_message << "save_file(): Couldn't save file: file failed at is_open() on the filename \"" << filename << "\"" << std::endl;
+      throw std::runtime_error(error_message.str());
+   }
+
+   return true;
+}
